Represent histogram values with integers.
Histogram values are represented with integers within FF, only the sum
is represented as long.

+++ src/main/scala/histograms/Histogram.scala
@@ -5,7 +5,7 @@ import org.json4s.jackson.JsonMethods._
 import scala.collection.mutable.{Map => MMap}
 import scala.io.Source
 
-case class RawHistogram(values: Map[String, Long], sum: Long)
+case class RawHistogram(values: Map[String, Int], sum: Long)
 
 sealed abstract class HistogramDefinition
 case class FlagHistogram(keyed: Boolean) extends HistogramDefinition
